Switch to Manage permission and future Jenkins compatibility (#126)
* bump to recent parent pom

* Compatability with Future Jenkins version

The code was incorrectly checking the permission ${it.ADMINISTER}
however it when vewing the root of Jenkins (and other views) is an
instance of AllView.  As such the permission check failed silently and
was then not shown.

https://github.com/jenkinsci/jenkins/pull/10678 sets ${it} to Jenkins
now, and then this causes the incorrect permission check to blow up as
it is a single permission that is set for permissions (which expects an
array).

This fixes the issue and actually removes the permission check from the
task, as the permission is already checked in the layout.

We now take advantage of the Manage permission also so that users
without Jenkins.ADMINISTER but with JENKINS.MANAGE can clear the build
queue.